Refactoring to handle Watch command type in the generic flow of command handlers (#1043)
