+++ internal/querymanager/query_manager.go
@@ -236,7 +236,7 @@ func (m *Manager) updateQueryCache(queryFingerprint string, event dstore.QueryWa
 }
 
 func (m *Manager) notifyClients(query *sql.DSQLQuery, clients *sync.Map, queryResult *[]sql.QueryResultRow) {
-	encodedResult := clientio.Encode(clientio.CreatePushResponse(sql.Qwatch, query.String(), *queryResult), false)
+	encodedResult := clientio.Encode(GenericWatchResponse(sql.Qwatch, query.String(), *queryResult), false)
 
 	clients.Range(func(clientKey, clientVal interface{}) bool {
 		// Identify the type of client and respond accordingly
