+++ internal/worker/cmd_meta.go
@@ -11,10 +11,24 @@ import (
 type CmdType int
 
 const (
+	// Global represents a command that applies globally across all shards or nodes.
+	// This type of command doesn't target a specific shard but affects the entire system.
 	Global CmdType = iota
+
+	// SingleShard represents a command that operates on a single shard.
+	// This command is scoped to execute on one specific shard, optimizing for shard-local operations.
 	SingleShard
+
+	// MultiShard represents a command that operates across multiple shards.
+	// This type of command spans more than one shard and may involve coordination between shards.
 	MultiShard
+
+	// Custom represents a command that is user-defined or has custom logic.
+	// This command type allows for flexibility in executing specific, non-standard operations.
 	Custom
+
+	// Watch represents a command that is used to monitor changes or events.
+	// This type of command listens for changes on specific keys or resources and responds accordingly.
 	Watch
 )
 
@@ -44,7 +58,7 @@ type CmdMeta struct {
 
 	// composeResponse is a function that combines multiple responses from the execution of commands
 	// into a single response object. It accepts a variadic parameter of EvalResponse objects
-	// and returns a unified response interface.
+	// and returns a unified response interface. It is used in the command type "MultiShard"
 	composeResponse func(responses ...eval.EvalResponse) interface{}
 }
 
@@ -53,12 +67,6 @@ var CommandsMeta = map[string]CmdMeta{
 	CmdPing: {
 		CmdType:              Global,
 		WorkerCommandHandler: eval.RespPING,
-	},
-	CmdAbort: {
-		CmdType: Custom,
-	},
-	CmdAuth: {
-		CmdType: Custom,
 	},
 
 	// Single-shard commands.
@@ -71,6 +79,16 @@ var CommandsMeta = map[string]CmdMeta{
 	CmdGetSet: {
 		CmdType: SingleShard,
 	},
+
+	// Custom commands.
+	CmdAbort: {
+		CmdType: Custom,
+	},
+	CmdAuth: {
+		CmdType: Custom,
+	},
+
+	// Watch commands
 	CmdGetWatch: {
 		CmdType: Watch,
 	},
